requirements.txt
amqp==5.1.1
asgiref==3.7.2
Babel==2.12.1
billiard==4.1.0
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-cors-headers==4.1.0
django-crontab==0.7.1
django-model-utils==4.3.1
django-money==3.1.0
djangorestframework==3.14.0
kombu==5.3.1
prompt-toolkit==3.0.38
psycopg2-binary==2.8.6
py-moneyed==2.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
